Carat section changes
Changed the Carat Management page layout and functionality.  Changed the layout for the way the carat section is displayed on the admin new/edit pages.

Changed the way the carat position select works - before this select dropdown defaulted to 25.  But, this caused problems if the user chose a number greater than the actual number of carats available.  So, this now does a count() of the available carat array and populates this select with this number.